Перенес фиксы из ветки annamaer, составил один хелпер

+++ src/s21_from_int_to_decimal.c
@@ -2,8 +2,8 @@
 
 int s21_from_int_to_decimal(int src, s21_decimal *dst) {
   reset_decimal(dst);
-  set_sign(dst, signbit(src));
-  if (signbit(src)) {
+  set_sign(dst, signbit((float)src));
+  if (signbit((float)src)) {
     src *= -1;
   }
   dst->bits[0] = src | BLANK;
